// cleveremote-front/src/app/services/collections/elements/interfaces/scheme.interfaces.ts
export enum SELECT_STATUS {
    SELECT = 'rgb(0, 150, 136)',
    WARNING = '#FFCD00',
    ERROR = '#FF0000',
    INPROCCESS = 'rgb(0, 0, 255)',
    INPROCCESS_BLINK = 'rgba(255, 255, 255, 0.01)',
    DEFAULT = 'rgba(255, 255, 255, 0)',
    OVER = 'rgb(139, 195, 74)',
    STOP = 'rgba(255, 255, 255, 0)'
  }

  export enum WORKING_STATUS {
    INPROCCESS = 'INPROCCESS',
    STOP = 'STOP'
  }